Add util decorators
